+++ tests/Extensions/NewRelicExtensionTest.php
@@ -8,13 +8,17 @@ class NewRelicExtensionTest extends TestCase
 {
 	public function testExtensionCanBeLoaded()
 	{
+		$this->assertCorrectArchitecture();
+
 		$this->assertTrue(extension_loaded('newrelic'));
 	}
 
 	public function testAgentIsStarted()
 	{
+		$this->assertCorrectArchitecture();
+
 		exec("pgrep -f /opt/newrelic/daemon.x64", $output);
-		$this->assertCount(2, $output, 'No running daemon processes found.');
+		$this->assertCount(2, $output, 'NewRelic daemon not running.');
 	}
 
 //	public function testDefaultEnvironmentValues()
@@ -34,4 +38,11 @@ class NewRelicExtensionTest extends TestCase
 //	{
 //		$this->markTestIncomplete('Not yet implemented.');
 //	}
+
+	private function assertCorrectArchitecture()
+	{
+		if (php_uname('m') !== 'x86_64') {
+			$this->markTestIncomplete("Testing NewRelic extension requires x86_64 architecture.");
+		}
+	}
 }